Restructured part of the pipeline, created controller class and renamed MainScript
